Split the Player Panel
Achieve #12644
